Add new valueEquals API for NavType
NavType.valueEquals can be overriden to compare equality of complex types. Defaults to this == other implementation.

Test: ./gradlew navigation:navigation-runtime:cC
Bug: 327229511
Relnote: "NavType has a new API valueEquals that takes two values of type T and returns true if the two values are equal to each other. This function can be overriden to check equality for complex types."
Change-Id: I6cb97c9b4a9992c5358f76c6940c2d9b78149f26