Update DragGestureNode to abstract touch slop skipping.
Touch slop skipping, or shouldStartDragImmediately is a boolean in all draggables. It is a lambda in DraggableNode because in Scrollable we need to call a function to get the most up to date information about skipping touch slop evaluation. This CL re-works that layering. ShouldStartDragImmediately changed to an abstract function in the DragGestureNode. This allows implementations of that node to lazily evaluate the need to skip touch slop. In a future CL, Scrollable can implement DragGestureNode directly and we can remove the need for a lambda in DraggableNode.

Test: Previous tests should pass.
Change-Id: I558c0423d691b93cceaf511115e08bba736e9663