Fix Macrobenchmarks for pixel watch target
ProgressIndicator and Expandables are failing on pixel watch target
since the content which is supposed to be clicked goes out of the screen
(due to the smaller size of screen).

For ProgressIndicator, the button which takes the test to the right
screen is out of the bounds of screen, which results into the button not
geting clicked.

For Expandables, when the first content is expanded, the second one goes
out of the screen and is no longer clickable.

Bug: 328016250
Test: Ran the tests on different emulators
Change-Id: Ibe92cca4d58e33c14ebc16e5e51489d0cbc51794