Add convenience APIs for Float, Int and Boolean in SQLiteStatement
Also add getColumnNames() to get all column names in a List, which means also updating gen code that used utility function and providing an overload in AmbiguousColumnResolver. Note that AmbiguousColumnResolver was incorrectly restricted to LIBRARY_GROUP but it is an API used in gen code, so should be considered 'pseudo public API' (i.e. LIBRARY_GROUP_PREFIX).

Bug: 322391411
Test: BaseConformanceTest
Relnote: "Added convenience APIs in SQLiteStatement for Float, Int and Boolean along with a getColumnNames()"
Change-Id: I50fd3bed69d5b1964b2f2d47f84b2b1b9c9ad4f6