[KSP2] Fix type parameter bounds
KSP2 adds `Any?` to the upper bound of type parameters when the bound is not explicitly specified. When the origin is Java this bound has a nullability of platform[1], which leads us to generate a Kotlin type name `Any` instead of `Any?` because `isMarkedNullable` is false for platform types. This CL forces the nullability of bounds to be nullable in KSP if it's of platform type. This also changes explicitly specified bounds, for example, `List<T extends Number>` in Java will become `List<T: Number?>` instead of `List<T: Number>` in Kotlin codegen.

If the bound is specified and has a `@NotNull` annotation the nullability will be correctly recognized as not-null and not platform in KSP. So `List<T extends @NotNull Number>` in Java will have a Kotlin typename of `List<T: Number>`.

Note that this change doesn't apply to platform types in other places. We may consider making them consistent in the future[2].

Also note that this changes the behavior in KSP1 as well, and the result is not consistent with Javac when the bound is explicitly specified. However, we don't do Kotlin codegen with Javac so it should be less of an issue.

[1] https://kotlinlang.org/docs/java-interop.html#null-safety-and-platform-types
[2] https://b.corp.google.com/issues/331427836

Test: manually tested with KSP2

Change-Id: Iccfd35a213c180f7f5e38cbd69146a1010c0b756