Fixes testModifiers_doNotCauseUnnecessaryRemeasure
This test was failing because the measurePolicy was being recreated on recomposition, causing a relayout. Before this change, the draw synchronization didn't wait for draw properly, so the assertion would run before the relayout and redraw finished. Now the test properly waits for draw to finish, and makes sure the measurePolicy isn't being recreated, so only the modifier implementations can affect relayout behavior here now.

Fixes: b/281171119
Test: SizeTest#testModifiers_doNotCauseUnnecessaryRemeasure
Change-Id: I955b3a9e6a3a0395abbd0428f086ebae4e61deda