Rename shouldShowKeyboardOnFocus and make nullable.
The old name was too verbose – the "should" is unnecessary, and we
even had some methods that omitted it.

This change also makes it nullable, to prepare for making
KeyboardOptions properly mergable.

These changes are all OK from an API compatibility standpoint because
this parameter was introduced in 1.7 alpha.

Bug: b/295951492
Test: n/a
Relnote: "Renamed `KeyboardOptions.shouldShowKeyboardOnFocus` to
    `showKeyboardOnFocus`."
Change-Id: Ib4b7af571c6bdfb06c3b0e160482cbb62c5277fa