Sort dackka args; publish samples for all (KMP) libraries
Now that we publish samples, we need these explicit dependencies for that to happen.

Bug: 329424152
We also need to enforce this; continuing in that bug.

The sorting is for reproducibility and allowing diffing the dackka args file; used in follow-up investigations
The conversion to a set is to remove any redundant elements.
Coverting to a List<File> is necessary because FileSystemLocation isn't Comparable

Test: enforcement in follow-up CL aosp/3006371
Change-Id: I8e9b5e201d5ac3a9ee426674163a866f0474393e