Fix Color, TextUnit and Offset savers
These savers have been causing crashes when AnnotatedString.Saver has been used. Specifically, needed to handle the TextUnit's and Color's Unspecified values.

Also updated the tests so that they test the actual code path. In particalar, we have a save and restore methods inside Savers class that are designed to save/restore nullable types. Specifically, null values saved as "false" value and then the "false" restored to the null value of the corresponding type.
However, the tests never checked this logic, so we actually never were able to restore the Offset.Unspecified correctly. Instead it was always restored as null and caused crashes.

Test: SaversTest
Test: demo with rememberSaveable of AnnotatedString
Change-Id: I0d0030a6f627cfc61f5d910398df759c26a1d594