Override build type props when build types already exist
It's currently possible to customize release build types creating
benchmark and nonMinified build types. For example to customize
`release` it's possible to create `benchmarkRelease` and
`nonMinifiedRelease`. This PR makes sure that even when they're
customized, properties that would break baseline profile generation are
overridden. This PR also makes sure that the signing config is copied
from the release build type to the extended ones.

Relnote: "when customizing a nonMinified or benchmark build type,
    ensure that properties that are needed to make baseline profile
    generation and benchmark working are overridden."
Test: ./gradlew :benchmark:benchmark-baseline-profile-gradle-plugin:test
Bug: 324837887
Change-Id: Ib8f0555001ff719f4b82c35cf50c13b4a3fa308d