Add ComposableNavHostFragment
Add a convenience extension on NavHostFragment
that automatically adds the
ComposableFragmentNavigator. This ensures that
developers can directly substitute
NavHostFragment for ComposableNavHostFragment in
a layout XML file.

Relnote: "Adds a new `navigation-fragment-compose`
artifact that includes a `ComposableNavHostFragment`
alternative to `NavHostFragment` that allows you to
add `composable` destinations to your Navigation XML
files. Each `composable` destination must be
expressed as a top-level, no argument `@Composable`
method whose fully qualified name is used as the
`android:name` attribute on each destination. When
navigating to one of these destinations, a containing
fragment is created to display the composable content."
BUG: 265480755
Test: new ComposableNavHostFragmentTest

Change-Id: I0ef2e5897128d18d1b374f3617c7095022ca7c0e