start video calls with the speaker endpoint; inc to alpha 4
A request came in to start video calls with the speaker route as the
active endpoint if earpiece is the current endpoint. This is reasonable
given that a video call with the earpiece as the active endpoint is an
unlikely scenario.

Relnote: "Redirect video call audio to speakerphone if the call
          initiates on the earpiece audio route"
Bug: 313971215
Test: manual + 4 end to end tests in
        `androidx.core.telecom.test.CallsManagerTest`
Change-Id: I5fa16879b054da17c9f92b49ed9810a0e7b7115a