Add LocalFragment to navigation-fragment-compose
Add a composition local that allows a composable
method inside a ComposableFragment to access the
Fragment instance.

This allows the method to extract arguments,
set transitions on the Fragment instance, and
otherwise avoid cases where you'd manually need
to construct a Fragment with a ComposeView as
its view just to access Fragment related properties.

Relnote: "The `navigation-fragment-compose` artifact
now provides a `LocalFragment` composition local to
composable methods within a `ComposableFragment`."
BUG: 265480755
Test: new ComposableFragmentTest added

Change-Id: If35e5405bbd94fd5f45c1cc0dfe3d74278d64a09