Clean up NavTypeConverter
Removed internal types for Maps, Sets and other types that do not have built-in support from NavType. This should clarify confusions on what types are natively supported by safe args and which types are not supported (aka required custom NavType).

These internal types should be re-added in the future if and when it gets native NavType support.

Test: TH
Bug: 188693139
Change-Id: Iabf112b941fd891c798ad451e4d298094978e4e4