[Extensions-Capture] camera-extensions outputs YUV_420_888 when CaptureProcessorImpl is enabled and convert it to JPEG in camera-core
Previously camera-extensions do the YUV to JPEG conversion but the problem is that it can't set the timestamp of the
JPEG Image and thus we aren't able to match the timestamp of the Image and the CaptureResult in ImageCapture in order to
know exactly which request does the Image belong to.

This timestamp match is critical to avoid the issues when interrupting a extensions takePicture actions.

This CL is to let camera-extensions outputs the YUV_420_888 image directly and let the ImageCapture pipeline handles the
YUV-To-JPEG conversaion and rotation and EXIF hanndling.

Bug: 318027789, 318451053
Test: extensions/ImageCaptureTest
Change-Id: I7fc225a0fab186d4af82764cf0fc23f00659a101