Fix text field being ignored by Talkback
Add state description to the text field when it's empty. In this case it won't be ignored by the Talkback which happens when the text field has a child which is not speakable.
In case of BTF2 it happens all the time because it provides magnifier semantics; in case of BTF1 this can also be easily "achieved" when adding a node with a testTag only (for example, an icon via the decoration box).

So when the TF merged config has no children with a text or a content description, it means that no label/hint or a placeholder was provided. So we add
a state description that the Talkback will use to realise it needs to focus on this text field.

Test: AndroidAccessibilityTest
Test: demo with Talkback, Switch access, Select to speak on
Change-Id: I3e0a5b989c20d3a15a8d13470bc328e63491a2e9