Don't let colors out of the valid range
A recent fix to Easing ca lead to colors being generated outside
of the valid range (0..1 for rgb in the common case, 0..1 for
alpha always) due to over/undershoot interpolators or numerical
imprecisions.

Being outside of the range is a problem for colors since the
Color constructor throws an exception in that case. We fix it
by clamping the values instead (which is more useful). We also
make color interpolation more robust in the first place by
clamping the interpolation fraction to the expected 0..1 range.

Test: ColorTest, AnimateXAsStateClockTest
Relnote: N/A
Change-Id: I698f054bfd2add4a2c8ea372653074a0c71ce5f0