Make connection and statament close() idempotent in native drivers
The SQLite docs mention that sqlite3_close should receive a pointer to a sqlite3 object that has not previously closed yet our implementation made it so that multiple calls to close() would invoke the native function. This CL uses the existing isClosed flag to avoid calling sqlite3_close on an object that was already marked to be destroyed. The SQLite docs don't have the same phrasing for finalizing (closing) a statement but this CL also updates the native impl just in case.

Bug: 318767828
Bug: 331622149
Test: Existing (but really can't repro bug)
Change-Id: Ib2d0b38307d462d3f7ec4d09dd49fa7b9cad7574