Opt-in Metalava with K2 UAST
This is an one-off case tracked at https://youtrack.jetbrains.com/issue/KT-61734
Meanwhile, the API with wildcard suppression was deprecated, and its
modeling is wrong in K1 in the sense that it added back wildcard.
Instead, by turning on K2 UAST, wildcard suppression is correctly
applied to deprecated function as well.

RelNote: N/A
Bug: 271369023
Test: ./gradlew checkApi
Change-Id: I6fdc69da4c9af78b2b55c39b6e79d094b4dc74b2