Fix PreviewViewFragmentTest#checkPreviewUpdatedAfterDestroyRecreate failure
Root cause: When onViewCreated is invoked, the view might not have been attached to the window yet. The view can't get the Display to obtain the rotation info at the moment.
Solution: Checks whether the view has been attached to the window or not. If it is not attached yet, add an OnAttachStateChangeListener to monitor the attached state to create the Preview.

Bug: 335314543
Test: run test on Firebase pixel2api28 and pixel2api30 emulators and lab devices
Change-Id: Ifc35e3983b2a1e239dc3ead3a375236a8edec66a