Graduate multiTouchSwipe
Graduate this API without further additions. While the documentation
said it would become possible to change the start and end times of each
pointer, this would require significant API changes. Given it is only a
niche case, it doesn't justify the changes. Note that it is still
possible to do that manually.

Bug: 261561237
Test: bOS
Relnote: "Move `TouchInjectionScope.multiTouchSwipe` from experimental
to non-experimental."

Change-Id: Ia08179629059db5b54ec6a335dbbc5736496c85b