Add hasRoute<T> API for NavDestination
Previous merge was reverted aosp/3047794. ExperimentalSafeArgsApi annotations are now removed from tests.

Convenience method to check if the NavDestination's route is based on KClass<T>.

Internally, the implementation is based on destination ID which was set during its initialization. Therefore if the ID was later overridden with a custom ID, the function will always return false.

Test: ./gradlew navigation:navigation-common:cC
Bug: 188693139
Relnote: "Added hasRoute<T> API for NavDestination to check if the NavDestination's route is based on KClass<T>."
Change-Id: I873b7d2beee8a589858e84f8e6f95f295c51e6b9

# Conflicts:
#	navigation/navigation-common/api/current.txt
#	navigation/navigation-common/api/restricted_current.txt
#	navigation/navigation-common/src/androidMain/kotlin/androidx/navigation/NavDestination.android.kt