Clamp cubic easing output values
With the new algorithm to solve cubic easing curves, we are at the mercy
of numerical imprecisions which can cause the transform method to return
a value that's off by 1 ulp, which may be undesirable in some cases.

This CL computes the vertical bounds of the cubic curve so we can clamp
the output value to the known valid range.

This CL also remove the clamp to 0..1 in PathEasing which prevents
easing curves from over/under-shooting, and cleans up a couple of things
(including reduce the i-cache pressure).

Fixes: 335640702
Relnote: N/A
Test: PathEasingTest, EasingTest (unit and instrumented)
Change-Id: I8212cfc95c4ee6e2b4e7cc9d84cda1dd10f845ce