Recreate GraphicsLayers after trim memory callbacks
If the application is backgrounded, it will eventually
receive a trim memory BACKGROUND call after 2 other
applications are launched before returning to the
original application. The trim memory callback will cause
HWUI to purge GPU resources including displaylists.
Update the draw calls for GraphicsLayer implementations to
verify if a displaylist is available first before rendering
and re-execute the recording lambda if a displaylist is not
available.

The trim memory callbacks also impact ImageReader. So recreate
the ImageReader and re-persist the existing layers. This helps
insure that GraphicsLayer displaylists would not get destroyed
when the application is in the foreground regardless if the layer
is not drawn for several frames.

While we cannot control when the OS decides to release GPU resources
we can at least ensure that the displaylists should not be discarded
when the application is in the foreground.

Verified with modifications to existing test cases to ensure
GraphicsLayers are rendered successfully in case the displaylist
is discarded by the system.
Tested manually with adb shell send-trim-memory because it
can invoking the shell command when the test process is in the
foreground is not allowed.

Fixes: 333584635
Test: Added test to AndroidGraphicsLayerTest and tested manually
with send-trim-memory adb commands

Change-Id: Id9f949cdc9305fe613ff016d8dc4b73614a0c28a