Add NavDestionation.hasRoute overload
Add an overload that takes a KClass<T> as parameter in the event where user only has access to the route kclass during runtime. Since KClass<T> cannot be passed as TypeParameter, we need an overload that takes a KClass.

Test: existing tests
Bug: 188693139
Relnote: "Add an overload that takes a KClass<T> as parameter in the case where users only have access to the route kclass during runtime."
Change-Id: Iea94de164795b79e99b8f1c2661e06ca6639b2bc