Generate binary compat stubs for nullable value classes
aosp/3017515 changed default parameter value class types to be nullable if inline class wraps a non-primitive value. Kotlin encodes nullability into the method mangle, however, and this breaks binary compatibility.

This change generates a wrapper function that simply redirects the call. This function will only be called if the module doesn't expect new mangle for the function and will be ignored normally.

Fixes: 335384193
Test: compiler tests + verified that repro project successfully runs with recompiled artifact.
Change-Id: I894651168d0521cdc639d4f4c415b28c4e5d348f