BTF2 fix long press starting from end padding
When selection is initiated via long press from an end padding on BTF2, dragging the pointer shows an incorrect selection, in other words changes the anchor point immediately after initiating a selection. This is due to the fact that `TextFieldSelectionState` chooses the anchor point after correcting a reversed selection.

Fix: 240624990
Test: gradle :compose:foundation:foundation:cAT
Change-Id: I102f978b02256c69989e9a11910040a3f4b298a9