addCall throws CallException out to client; bump to alpha 07
Previously, when the platform failed to add the new call request,
Core-Telecom would throw a CancellationException out to the client. That
method failed to take advantage of the CallException reason the platform
provided.  Now, Core-Telecom will provide an error code for why the call
could not be added (e.g. the ongoing call is unholdable).  This is
helpful so the client can provide UI or actions related to the add call
failure.

Relnote: "Core-Telecom will now throw a CallException if the platform
         failed to add the call due to an exception, reaching max call
	 count, etc."
Fixes: 335662722
Test: manual + unit tests + e2e tests
Change-Id: I41f27f24423e0c735f2ff2c4d203a07674877b9b