Include class name in saveable delegate key.
Including the class name of the property owner as key for the saveable avoids conflict with same name properties in different classes using the same SavedStateHandle from the same ViewModel. Note that 'thisRef' (owner) is not always available, for example, when the delegate is used in a local property of a function.

Bug: 331695354
Test: SavedStateHandleSaverTest
Change-Id: Ifec9c9c6278e49f54cf41fac3c48d527d5bf23e4