Remove an allocation from withFrameNanos()
The lateinit var being allocated inside a lambda was causing
the inliner to miss an important optimization and to wrap
the var inside an ObjectRef. In our case we always want to
allocate the FrameAwaiter anyway. The only drawback is that
if there's a failureCause, we will have allocated the awaiter
for no reason, but it doesn't matter in that case.

Test: BroadcastFrameClockTest
Relnote: N/A
Change-Id: Id8dfc6e152b84b57d7a81cfa7a61271db3335cdf