Remove suspend callbacks from draggable2D
The suspend callbacks in draggable2D should be removed in favor of simple callbacks. This removes the implicit scope that was provided in these callbacks where the lifecycle was attached to the Draggable2D modifier. This change allows the lifecycle of the events called in these callbacks to be better expressed and controlled by the caller.

Test: N/A
Relnote: Replaced onDragStarted and onDragStopped with non-suspend callbacks.
Change-Id: I59de8c2b7adb0353420da1e7c8f3c9aa9d46f7e7