[KSP2] Keep the variance of component type in getArrayType()
In KSP2 varargs are no longer of array type and we rely on getArrayType() to convert it back to array type to keep the behavior consistent between Javac/KAPT and KSP. However, our getArrayType() ignores the variance of the passed in component type and makes it always invariant. This CL adds the variance back.

Test: XArrayTypeTest.kt
Change-Id: I773186436f127fded3af10defe188c7ce031d252