Explicitly measure child in SandboxedSdkView
During the measure pass of SandboxedSdkView, explictly call
measure() on the child view, if one exists. This will ensure that
if the local view is a ViewGroup, that it will get the opportunity
to measure its children so that they are laid out correctly.

Added a test to verify that the child of a FrameLayout that is returned
as part of a SandboxedUiAdapter.Session is laid out.

Relnote: "Fixed a bug that caused rendering issues for ViewGroups in
    backwards compatibility mode"
Test: IntegrationTests, manual test with local adapter and FrameLayout
Bug: 333849623
Change-Id: I8de92044eee18d7aee96cb04e62e7d1ca89b2a12