Fix for lazy list items displaying nothing during the disappearance animation
It was happening when the item content was scheduled to redraw (because of the click indication). The LayoutNode with this item is then immediately deactivated, however the content will still be redrawn as it was invalidated already. Now we will be releasing the layers when the node is deactivated in order to make sure it will not be redrawn anymore so we can continue displaying the content.

Test: new test in DrawingPrebuiltGraphicsLayerTest
Fixes: 339120677
Change-Id: I6cebe231cb61756afa7520bbd9239e2c352d9c36