Wait for window focus in CustomEffectContextTest
Sometimes the window is not yet focused when the test is started. When
the window does get focused, this leads to an state change at a
non-deterministic time and can change the expected order of execution.
Explicitly wait for window focus at the start of this test.

For varies reasons this cannot be applied generically to the test
harness, mostly importantly because there are valid use cases where
we expect window focus to go somewhere else.

Fix: 328299124
Test: Ran CustomEffectContextTest 200x
Change-Id: Ibb0fb0c8b8b40a0519c9ad20101286edf903e8b9