Remove ExperimentalWindowApi annotations from window APIs.
This is based on the annotated APIs from window/window/api/current.txt. Most of the experimental annotations have been removed except for the annotation for WindowAreaSessionPresenter since WindowAreaSessionPresenter.getWindow() is not yet in the public API window/window/api/restricted_1.3.0-beta02.txt. As a result, the experimental annotations for WindowAreaController, WindowAreaInfo, WindowAreaInfo.Type, and WindowAreaPresentationSessionCallback also cannot be removed yet. This also excludes all APIs under `area`, WindowMetrics.getWindowInsets(), and EmbeddingAdapter, which are still experimental.

Bug: 333837852
Test: N/A
Relnote: N/A
Change-Id: I101e6ad32dfc0dbca3adf90b45eb46e327f4ac9b