Fix baseline profiles and benchmarks on android multiuser
Benchmark and Baseline profile generation require transfering files
between storage accessible by shell and storage accessible by the user.
Shell runs on user 0. When the android selected user is also zero, shell
can see both storages. When the android selected user is not zero, shell
cannot see user storage.

As a result we cannot invoke shell commands operating in user storage
to copy and move files. Instead we need to utilize the running
instrumentation process to access user storage and shell to access shell
storage. In order to communicate between these two, the class
VirtualFile implements a number of methods to copy streams between the
two. When in shell storage we can use commands like `cat` and `cp` from
`/dev/stdin` to a file to write. When in user storage we use the java
api of the running process.

Test: Added and manually. Note we'll need to enable multiuser devices in CI.
Bug: 356684617
Bug: 373641155
Relnote: (Experimental) Enable Baseline Profile generation, and benchmarking on apps installed to a secondary user, such as all apps on headless Android Auto devices. This support has been tested in some scenarios, but let us know with a bug if it doesn't work for you.
Change-Id: I9fcbecf2a9c9075d161de69118c443d7a37102c0