Fix focus restoration when item goes out of bounds
We use the composition hash to restore focus. If the item to be restored goes
out of bounds and is disposed, it can't be restored. This CL pins the item
that is saved for focus restoration so that it is not disposed, and is
present when we are restoring it.

Bug: 330696779
Test: Added a new test to cover this case.
Change-Id: Ia51d19d790978202ac2c2551015881077d0c8ca6