Fix crash or incorrect compsition at end of nodes
If a group was removed at the end of a node group the
composer could renerate the remove operation to the wrong
node tree level.

This was caused by the coalescing of removes and
tree navigation not being correctly guarded that would
cause the move up the tree to be issued before the
remove which could either remove the wrong node or
cause a crash in the applier where it tries to remove
a non-existent node.

Test: ./gradlew :compose:r:r:tDUT
Fixes:  339618126

Change-Id: I9a13f10b9d4126abf2d12a91a2b1796b7f188f35