Extensions-Capture] Fix the onCaptureCompleted / getAvailableCaptureResultKeys issue in extensions testlib
The new extensions capture pipeline will have the following expectation for the OEM
implementations (which was tested in CTS)
1) if getAvailableCaptureResultKeys returns a non-empty list, then
   either SessionProcessorImpl.CaptureCallback#onCaptureCompleted (advanced) or
   ProcessResultImpl#onCaptureCompleted (basic) will always be invoked.

2) the onCaptureCompleted will have the same timestamp as the Image.

This CL is to align the extensions testlib to the above expectations.
Besides,
For advanced extender, it extends the LongCaptureAdvancedExtenderImpl
to the ConfiguratableAdvancedExtenderImpl and allow it to configure
the willInvokeOnCaptureCompleted and longDurationCapture parameters.

Bug: 339762546
Test: manual
Change-Id: I280e7a2ce96b23210185db70c2ea65f7cefb5a29