New API: Feature combination query
The API returns a result that combining Google Play Service data and the result from Android framework.

- CameraDeviceSetupCompat: a compatible version of CameraDevice.CameraDeviceSetup. It's an interface so it can be implemented with different data sources: camera2, Play Services, etc.

- CameraDeviceSetupCompatProvider: a interface that provides a CameraDeviceSetupCompat. We need this abstraction to cache it because getting the provider is usually costly.

- CameraDeviceSetupCompatFactory#getCameraDeviceSetupCompat: a compatible version of CameraManager#getCameraDeviceSetup. It returns a CameraDeviceSetupCompat.

Relnote: an API for querying supported camera feature combination. For example, checking if HDR and 60FPS can be enabled at the same time. This is a compatible version of camera2's CameraManager#getCameraDeviceSetup#isSessionConfigurationSupported API.

Bug: 336326522
Test: manual test and ./gradlew bOS
Change-Id: I3e927bf717d659eb283aee5af9e039e9061fd7f8