Changes compositingStrategy for SLC
The default compositingStrategy when changing alpha performs rendering to an offscreen buffer, however this in practice also means clipping content to the wrapper's constraints, which is sometimes not desireable (e.g. when using SwipeToReveal). Using an offscreen buffer is also slightly less efficient, so this change slightly improves the performance as well.

Change-Id: I995ffb2e7e72b6208dc303ea33c72b7e8dfaa152
Test: ScrollBenchmark macrobenchmark
BUG: 329702048