Reuse layer objects
It allows us to save on allocations and to do less work when we need to create a new layer. The caching mechanism is the same as was used in the previous implementation and is based on weak references.

Benchmark results.
LazyListScrollingBenchmark.scrollProgrammatically_newItemComposed
Before: 7,241,042 ns, 1162 allocs
After: 7,009,125 ns, 980 allocs
LazyBoxWithConstraintsScrollBenchmark. frameDurationCpuMs
Before: P50   24.7,   P90   34.9,   P95   37.2,   P99   43.2
After: P50   24.6,   P90   33.9,   P95   35.8,   P99   39.3

Bug: 340651308
Test: run benchmarks, existing tests in all modules
Change-Id: I62f53a1a17ca874fad6241a17a3fdf30845c5531