AuthenticationResult IntDef and Constructor Change
This adds, to the AuthenticationResult, the proper IntDef for the Jetpack
error codes. This also matches the constructor change that was
requested, where the framework constructor becomes internal, and the
Jetpack constructor is exposed.

Note that due to how Kotlin works with overloaded constructors, we
utilize a similar strategy to the 'fromBundle/toBundle' approach and
generate a boolean value-param that allows us to do 3 things:

1. Since both framework/jetpack need a single 'int', this ensures we
   don't face an overloads issue for 2 constructors.
2. Since we want to keep things immutable, this ensures the codes, when
   constructed, can be converted to their immutable quantity.
3. While a necessity to avoid the overloaded constructors, we also
   gain a parameter that allows us to log if a conversion happened or not.

Bug: 340911451 340912050
Test: JUnit Test Included

Change-Id: I20cb85fb26c3eeb69a7bc4f2dc055bf8f52547b0