[M3] Add ScrollIndicator to Material 3 for WearOS
This ScrollIndicator features a new design. It also fixes the problem of the jumping size- it's fixed now and depends on initial position of the List.
Bug:286883009
Test: "ScrollIndicatorTest.kt"
Relnote: "Add ScrollIndicator to Material 3. This ScrollIndicator features a new design. It also fixes the problem of the jumping size- it's fixed now and depends on initial position of the List"

Change-Id: Ic228d48ed349dabb9c1fa02b949f07e3a6a4781b