Fix the issue that wrong device id is used in virtual device
In virutal device manager environment,
when activity is running in both host and remote devices,
Its application context will have a device id where latest
activity was launched.

for example,
if activity context has a default device id, it is possible
that getApplicationContext() has a virtual device id. This is
incorrect and we should set the applicationContxt's device id
to default device id as well.

Bug: 339803490
Test: ContextUtilTest, virtual device environment manual test

Change-Id: Ic4aa29f165460c3a5375ac76a6897be32f57aad1