Eliminate an unnecessary copy in renderMultiBufferedLayer
Before this change, the Collection of params passed to renderMultiBufferedLayer was copied into an ArrayList unless params was already a MutableCollection. This is because mSegments contains MutableCollection, dating all the way back to the original implementation of this code. But the contents of these collections are no longer constructed one item at a time. So this is currently doing a copy for no good reason. (We probably don't want defensive copies in code focused on low-latency graphics like this. The API doesn't suggest that it copies. And if we were doing defensive copies, we wouldn't want to specifically avoid copying MutableCollection.)

Test: Existing unit tests
Change-Id: I53c595cc45fba03cc0114633266d4855ee33552a