Make sure to reset touch mode in tests
Some tests don't properly reset touch mode after running, which means that the device can be left in an unintended state.

Because FocusTargetAttachDetach test had this problem, it meant that some other focus tests in ui were accidentally passing in CI because it was being ran in keyboard mode, but failing when ran individually / not as part of the whole suite. This CL also ignores these broken test accordingly.

Bug: b/325466015
Test: tests
Change-Id: I6fed3359880fe0def64f482647250862e618909e