All Entry Subclasses Open BioPromptData
This adds the BioPromptData to all the Create and Credential Entry subclasses,
namely the PublicKeyCredentialEntry, the CustomCredentialEntry, and the
PasswordCredentialEntry for CredentialEntry, and CreateEntry on its own
for the creation perspective. Thus providers read the below from their POV
upon usage:

---------------------------------------------------------------------

@property biometricPromptData the data that is set by you to optionally utilize a credential
manager flow that directly handles the biometric verification and presents back the
response; set to null by default

---------------------------------------------------------------------

The bit above can now be utilized by providers.

Bug:  339650358
Test: Unit Tests already present
Relnote: "Adds the BiometricPromptData functionality to the
PublicKeyCredentialEntry, PasswordEntry, and CustomCredentialEntry, the
subclasses of CredentialEntry, as well as to the CreateEntry, so providers can opt to use the
Credential Manager Biometric Prompt flow."

Change-Id: I8e5bc104b7337850ca6c49064502044d29c1308b