Handle null values for built in CollectionNavType
Previously, null string lists (i.e. NavType.StringListType) are treated as a listOf("null") when it should just be null. On the other hand, other primitiv
e lists (i.e. NavType.IntListType) are treated as if there is a lack of value and the null value is not added to argument bundle.

Now, both lists with null values are added to the bundle as bundleOf("argName", null). This means NavBackStackEntry.toRoute() will also be able to decode
the deeplink back into a Route with a null arg value.

However this comes with a matchDeepLink behavior change. In the past, a non-String List-type query arg of value "{argName}" is treated the same as null, whereby both are treated as if lack of value, and neither arg values are added to the bundle. So both these routes are essentially interchangable:

"www.test.com?argName={argName}"
"www.test.com?argName=null"

But now with null considered as a legitimate value and added to the bundle, they are no longer interchangable.

Test: ./gradlew navigation:navigation-runtime:cC
Test: ./graldew navigation:navigation-common:cC
Bug: 342672856
Relnote: "Navigation can now handle navigating with a null value for the built-in nullable CollectionNavTypes"

Change-Id: I59d308e76045f8734da4e2fc2913148b1910f194