Refactor NavControllerRouteTest
Several tests currently assert that navigaton treats a query argument of "{argName}" the same as null. But they are different values - the former is a lack of value, while the latter is a value of null.

Refactor tests to assert {argName} against routes with {argName}, and assert null against routes with null.

Test: ./gradlew navigation:navigation-runtime:cC
Bug: 342672856
Change-Id: I255269084b48fb1d0fac455939d329e5a40c33fd