Fix ProcessCameraProvider documentation.
Kotlin doesn't support @inheritDoc and will automatically inherit the
doc if the overrided method doesn't have any documentation. So remove
the annotation.

Also remove the redundant ducumentation from the overrided methods that
have identical documentation as its parent.

Move the comment in the sample that was showed in the document incorrectly.

Bug: N/A
Test: ./gradlew camera:camera-lifecycle:build
Change-Id: I85e3d2be2ba55b68cf0ad9c70a7f8871faa3b0bc