Use frame time from Choreographer callback in prefetch.
Previously we were using view.drawingTime as the time when the current frame started, as it was done in RecyclerView. However in frames when there were no redrawing this property will not be updated, which results in the prefetch logic assuming there is no time available in the frame. Instead we can just use the value representing the frame start time we get directly from the Choreographer frame callback. View.drawingTime is using the same value in the frames, where there is a drawing happening: https://cs.android.com/android/platform/superproject/main/+/main:frameworks/base/core/java/android/view/ViewRootImpl.java;l=5083;drc=cab016098de9d7ebc4444374b4243747e575f3a6.

This change slightly reduced frame jank percent of ComplexNestedListsScrollBenchmark: from 55.3 to 52.13. However the effect will be even more noticeable when someone is trying to write a custom prefetch strategy, scheduling work in frames, when we don't normally draw. This issue was found by the benchmark in Play store where exactly this was happening. In our default implementation we usually prefetch during fling, which means there is a drawing happening in those frames as well.

Test: run ComplexNestedListsScrollBenchmark
Change-Id: I375075c737f9a22b3c391bc5da7a9323c2dedfb6