Recreate lazy layout's info object on small scrolls
Previously we were reusing the same info object in such cases, however it proved to be hard to use when the info observation is done from snapshotFlow { } as it automatically skips the updates if the previous and old objects are equals (true here as the object didn't change). Users had to map the info into their own immutable objects to make it work. Now we will be recreating the object instead for making this common use case simpler.

Fixes: 316936479
Test: new tests for all the lazy layouts
Change-Id: I1742b49525acae3f4d1dc1f698db42f3754ef40e