Resolve invalid Relation codegen.
Users were encountering a "missing recursiveFetchLongSparseArray()" error in
codegen for iOS. This issue was due to the fact that room-runtime was not
depending on androidx.collection in common. Now that androidx.collection is
available in KMP (and, therefore, LongSparseArray is available) we can
depend on it in room-runtime.

This issue was reproduced and verified by adding the androidx.collection
dependency to the common source set of the kotlin multiplatform test app
and using existing tests that use @Relation on the iOS target. Once verified, the
dependency was added to room-runtime common source sets, and the
`recursiveFetchLongSparseArray()` function has been moved from the
Android-only RelationUtil.android.kt to the common RelationUtil.kt.

Bug: 341002184
Test: BaseQueryTest.kt
Change-Id: Id1afde6c5eeb46c3fe0bf9cc14c52eccb161be72