Add explicit setter for Row.onClickDelegate
The new setter is restricted to library & host code for now, because 3p
should generally not be constructing delegates manually.

This setter is needed within the host, for a feature that transforms
GridItems into Rows.

More generally, Delegates must be exposed in our public API to facilitate
copying & transforming client-side data objects. We have discussed and agreed
on publicizing Delegates previously, when we encountered the need to
instantiate Delegates for Google-only templates.

Bug: 310757799
Test: ./gradlew :car:app:app:test
Relnote: N/A
Change-Id: Ia187be8872f6537c1da78f19b15123799425a9b9