Improve modifier set access in snapshots
Now passes the `modified` set as a parameter to helper
functions during `apply()` of a snapshot to avoid a case
where the `modified` set could get out of sync with the
set read outside the synchronized block and could have
caused the bug reported in 344498074.

No new test was added as this crash was not reproduced
locally.

BUG: 344498074
Test: ./gradlew :compose:r:r:tDUT
Change-Id: Ie65a486116f30d60b48fdd89dfca5b6d84405e0c