Ensure RenderNode#endRecording is always called
Mirror View behavior in that each attempt to draw
in a RecordingCanvas obtained by RenderNode#endRecording
is wrapped in a try/finally block. This defensive
check ensures that regardless the RenderNode can still
be used regardless if recording failed. This is important
to ensure RenderNode instances are in a consistent state
to support recycling behavior.

Bug: 343750859
Test: Added test to [Android|Desktop]GraphicsLayerTest
Change-Id: I3caae0d61a9e6e04591b9ca1a27e960292599fb9