Add a lint check to move type-use array annotations
This is intended to be used once per project as they move to a newer version of the annotation library with type-use nullness annotations. It is intentionally not added to the issue registry because it shouldn't be run right now.

It is needed because in Java, a type-use annotation preceding an arary applies to the array element type, not the array itself, so just switching annotations to type-use without moving the annotations on existing arrays will change the meaning of the API surface. See https://docs.google.com/document/d/1tNqV9Dgtii6wvulm8emjs0bwgSL3Fwf3F4LFXfVeyTU/edit?pli=1&tab=t.0#bookmark=id.ievd7f7tcqf1 for more details.

I tested this on the :core:core project with the following steps:
1. Added the new issue to the `AndroidXIssueRegistry`
2. Ran `./gradlew :core:core:lintFix`
3. Added `TYPE_USE` target to `androidx.annotation.NonNull/Nullable`
4. Updated `core/core/build.gradle` to a project dependecy on `:annotation:annotation`
5. Ran `./gradlew :core:core:updateApi`

I manually checked some of the modified files from step 2 for formatting, and verified that step 5 did not modify any API files (so the API surface stayed the same).

Bug: 338099525
Test: ArrayNullnessMigrationTest, plus testing on the core library as described above
Change-Id: I8ac22dd1aa79e6a1b80449768ba80e39bc3838a4