Forward fix null lists
We merged aosp/3105059 to fix null lists by adding null values to the bundles.

In this forward fix, we handle null the same as a missing value for lists whereby we will default to the defaultValue if present, otherwise we default to an emptyList().

Test: ./gradlew navigation:navigation-runtime:cC
Test: ./graldew navigation:navigation-common:cC
Bug: 342672856
Relnote: "Navigation handles null value for built-in CollectionNavTypes by by defaulting to the NavType’s defaultValue, or an emptyList() if there is no defaultValue."
Change-Id: I84158320503bc04052de8f80022f83d86e23cc15