Refactor generateRouteWithArgs
Add new internal KSerializer<T>.forEachIndexed() function and applies an operation lambda on each child argument of the serializer. This function takes a Map<String, NavType<Any?>> and provides this navType into the operation lambda.

Refactored RouteEncoder to output a map of argName to List<String> (the value is the output of NavType.serializeAsValue/serializeAsValues).

Finally refactored generateRouteWithArgs to use the new forEachIndexed(), new RouteBuilder, and new RouteEncoder.

Test: existing tests
Bug: 340966212
Change-Id: Ic2ebe6cc7c396254c10a03c26639432d9c6c6fe3