Add code folding hints to make the demo app manifest file more readable.
This CL adds code folding hints for Android Studio and
reorganizes the demo app manifest file to make it more readable.
It groups activities and services by functionality and
adds region...endregion comments:
https://www.jetbrains.com/help/idea/code-folding-settings.html
to each code folding section to explain the purpose of the section.
This makes it easier to find the relevant activity or service and
to understand their purpose.

Fix: 346716118
Test: Check collapse view with Android Studio: http://screen/6voZBoVftdVzEq5.png
Change-Id: Ifdf2310559b7eac5223316a5a5ca741ee873226c