Correctly calculate consumed scroll with beforeContentPadding
In some cases (when we start scrolling from non first item, and scroll delta is larger than we have to scroll, and we have non-zero beforeContentPadding) we were incorrectly calculating the scrolled delta amount. As a side effect it was causing unexpected placement animations when Modifier.animateItem() was set, however the real issue is in how we calculated the consumed scroll delta. The same issue was fixed in lists, grids and staggered grids. Pager has no notion of "consumed scroll delta", so it wasn't affected (in Pager we pre-calculate how much scroll we would consume as we know the page sizes in advance).

Test: new test for list, grid and staggered grid
Fixes: 340194454
Change-Id: Ie6f928d7ce5777464d641a50565e7d0a21256dff